DIM Test Refactor (#994)
### Summary
<!-- Quick summary of changes, optional -->
Experimenting with separating tests into different files by function.
This is pretty pointless on the DIM, as there aren't many tests, but for
boards such as the BMS which have many tests that end up all being
crammed into Test_StateMachine.cpp, having a better mechanism to
separate tests into distinct files is valuable. (See Test_Faults.cpp and
Test_Eeeprom.cpp for example, they require the same setup as the state
machine test, and it means that 3 files need to be maintained any time
the state machine test setup or teardown needs to be modified.)

Admittedly the DIM is a poor example as to why this is beneficial, and I
would argue that tests have been over-segregated in this example (just
to get the point across), however, I cannot re-iterate how much this
would benefit other boards.

#### Changes
* Move `DimStateMachineTest` class definition to a header file
accessible to any .cpp test file. Also renamed to
`DimBaseStateMachineTest` to fall in line with the existing
`BaseStateMachineTest` class.
* I'm leaning towards naming it something more along the lines of
DimBaseTest, as even code that does not require the state machine
running can be tested within it (albeit slightly less efficiently, but
none of our tests are that demanding anyway and will allow for a single
source of truth when it comes to test initialization)
* `test_stateMachine.cpp` should focus on testing state transitions in
my eyes (open to some other things residing here as well, please share
feedback), and should not be the defacto place to add a new test that is
unrelated to state machine behavior, instead create a new test class
inheriting from `DimBaseStateMachineTest` and group tests with similar
functions/goals into smaller test files.
* Because the DIM has no state transitions, there is no longer a
`test_stateMachine.cpp` file


### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->
- test_stateMachine.cpp split into multiple files.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- All existing tests pass

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).

---------

Co-authored-by: Gus Tahara-Edmonds <gtaharaedmonds@gmail.com>