Remove -werror flag (#1015)
### Summary

Remove `-werror` flag, which causes all warnings to be treated as build
errors. This hasn't been a problem in the past since everyone has been
using the same version of `gcc` and `MinGW`, but `clang` on macOS is a
bit more picky and tests builds are not compiling on some Macs.

We could either 1) go in and fix all warnings or 2) remove this flag and
permit warnings. some of the warnings are related to autogenerated code
and are so nontrivial fixes.

Since Mac users are going to be working on tests in the next few days
they need builds to be passing - I think we should just remove this
flag. This is just for the test binaries so pretty low risk.