Latching faults and warnings (#1024)
### Summary
<!-- Quick summary of changes, optional -->

Faults and warnings (finally) latch. This is implemented in another CAN
message for faults and warnings, which counts the number of times an
alert is set. Each count gets 3 bits for a maximum count of 7. This also
means the maximum number of faults or warnings is reduced from 64 to 21,
since currently we just pack all faults/warnings counts into a single
CAN message and so can only fit 21 counts for a total of 63 bits.

So:

- `<board>_Faults` is unchanged (still describes the current status of
each fault
- `<board>_FaultCounts` allocates a 3-bit counter for each fault, which
count the number of rising edge transitions (no fault -> fault)

(and similarly for warnings)

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Implement fault/warning counts as described above and fix tests
- Change the alert settings syntax to be more similar to
`App_CanTx`/`App_CanRx`

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Spun wheels on jacks

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).