Faults cause exit DCM drive state (not opening contactors) (#1025)
### Summary
<!-- Quick summary of changes, optional -->

The current fault handling behavior is to open contactors. This is super
annoying (and would mean the end of competition) since the BMS latches
whenever contactors open, and so needs a power cycle.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Modified state machines so:

- If the BMS detects a fault, it doesn't do anything
- When the DCM detects a fault during drive state, it goes back to init
state
- Init -> drive transition on the DCM is blocked if there is a fault

By faults here I mean the faults defined in the `*_alerts.json`.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

TBD, I need to think about this

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).