Upgrade STM32CubeMX version from 5.3.0 -> 6.9.2 (#1016)
### Summary
<!-- Quick summary of changes, optional -->

Upgrade the boards in our repo from STM32CubeMX 5.3.0 to 6.9.2. We need
to switch to 6.9.2 to unlock support for the STM32H7xx family of
microcontrollers, which is going to be on our vehicle controller this
year. Also, we should just upgrade since they probably fixed a bunch of
bugs.

TODO: Test this on other people's machines

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Migrate every board's `.ioc` file from 5.3.0 to 6.9.2. This means we
had to regenerate the drivers/middleware for every board, leading to a
crazy diff.
- Switch to CMSIS-RTOS v2, which is supported in 6.9.2 (it wasn't in
5.3.0). CMSIS-RTOS is an API that wraps FreeRTOS, and v2 is way easier
to use than regular FreeRTOS and way more powerful than v1.
- Bumped the version in the container (using a custom tag currently so
`ubcformulaelectric/environment` still points to the regular container,
and other CI jobs can run).

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Flashed all boards and spun wheels on jacks
- [x] This is a big change, so lets do some driving with these changes

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).