Untrack generated STM32Cube drivers and middleware (#1040)
### Summary
<!-- Quick summary of changes, optional -->

We currently use STM32CubeMX to generate our HAL drivers and FreeRTOS
independently for every board. This is dumb because we track hundreds of
thousands of lines per board that we don't write ourselves. These HAL
and FreeRTOS libraries are available as a GitHub repo, so we should
subrepo them instead.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Subrepo STM32CubeMX F4 firmware package
- Significantly refactor CMake to build libraries instead of passing
around massive lists of source files
- Also just deleted the GSM since its currently an F3 project but will
be made with an F4, so there's really nothing there worth keeping around

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Get some driving time (this is a massive change, so we should take
in on track)

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

SW-43

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).