Prefix alerts with their type (#1061)
<!-- Quick summary of changes, optional -->

When testing the new faults and warnings on the car, it was very
annoying to diagnose shutdown since the new alerts are not prefixed with
`Fault` or `Warning`. Changed it so they are automatically now.

<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Prefix faults/warnings with `Fault` and `Warning` automatically.

<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] CI builds

<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).