Platform Independent JSONCAN (#1058)
### Summary
<!-- Quick summary of changes, optional -->

Make `jsoncan` platform independent. This is required so @Lucien950 can
use it on the dashboard, without our STM32 firmware.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

To use `jsoncan` on any platform, you need to:
- Define the flag `THREAD_SAFE_CAN_PACKING` if you don't want to use the
FreeRTOS thread safe guards
- Define a file called `Io_SharedCanMsg.h` that defines the struct: 
```c
struct CanMsg
{
    uint32_t std_id;
    uint32_t dlc;
    uint8_t  data[8];
};
```
- Call `Io_CanTx_Init` and pass the transmit function that actually
sends the message (or pushes it to a queue to be sent later). The
function returns void and takes a pointer to a `const struct CanMsg`.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

None yet.

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).