Enable HardFault handler on all boards (#1055)
### Summary
<!-- Quick summary of changes, optional -->

On ARM cores, when you do something very very bad an interrupt is fired
which is handled by each board's `HardFault_Handler`. When doing so, the
core's registers are pushed to the stack. The previous leads have
written some black magic to get these values viewable via a hooked up
debugger. I enabled this handler callback for all boards, because it was
previously only enabled on the BMS and DCM.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Add hard fault handler to board `stm32f4xx_it.c` files.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

N/A, but this is only changing the PDM, FSM, and DIM with code that has
been on the BMS and DCM for months. Plus, this only caused a change in
behavior after a critical fault, so the risk is low.

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).