Track .ioc checksums to remove code generation from CI (#1087)
### Summary
<!-- Quick summary of changes, optional -->

We currently re-generate STM32CubeMX code during CI, which is flaky and
slow. This is because the binary builds depend on a checksum of the
`.ioc` files, which are generated during STM32CubeMX code generation.
These are currently git ignored, forcing code regeneration in CI.

However, we can instead track the checksums, and add a new job to verify
the `.ioc` matches the tracked checksum. This gives us the best of both
worlds:
- We can verify that the `.ioc` file is up to date with the code
generated by STM32CubeMX by comparing the checksums. If they don't
match, you will need to re-generate code locally.
- CI is fast.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Track `.ioc` MD5 checksums
- Add H7 dev board (I forgot to do this before whoops)

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).