CAN Bootloader (#1083)
### Summary
<!-- Quick summary of changes, optional -->

Create a CAN bootloader for STM32F412 targets so we can update their
firmware over the CAN bus. A bootloader is a separate program running on
the MCU that (unsurprisingly) handles:
- Booting: Can start the main program (usually called the "app", not to
be confused with our app-level abstraction layer)
- Loading: Can update its app code

When you reset the microcontroller, instead of running the usual program
it now runs the bootloader. Its then up to the bootloader to decide
whether or not to start the app or update its code. A microcontroller is
able to overwrite its own flash: So its only a problem of telling the
microcontroller what data to write at what location.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Create STM32F4 CAN bootloader in `firmware/boot` directory. The
bootloader handles updating the application code, and uses STM's cyclic
redundancy check peripheral to verify that the application was
programmed correctly. Also, it broadcasts its application state at 10Hz
over CAN while in the bootloader.
- Created linker scripts for the app and bootloader in `firmware/linker`
to place the programs in their respective locations in memory.
- Add Python script to parse and send the bootloader a binary over CAN.
Unfortunately this has to be a separate repo, since WSL has no CAN
support whatsoever. This can be used to update or erase boards'
applications.
- Add shared `hw` drivers for FLASH memory and cyclic redundancy check
peripherals

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- [x] Testing on all boards

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).