Refactor BMS (#1063)
<!-- Quick summary of changes, optional -->
Refactor the BMS to be more in line with the new coding structure (like
we have on the DIM)

<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->
BMS app and IO layer follows the new structure.

<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->
Tested while accumulator was partly dissasembled, things not yet tested
include:
- Charging
- Spin Wheels
- TS Voltage Sense
- TS Current Sense
- Pre-Charge

BMS code has been modified to not allow it to enter Precharge state for
any reason.

<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).

---------

Co-authored-by: will-chaba <willchabaubc@gmail.com>