Bootloader on H7 (#1105)
### Summary
<!-- Quick summary of changes, optional -->

Port our CAN bootloader to the STM32H733xx microcontroller, so it can be
used with the BMS and VC on Quadruna.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- STM32H7 bootloader
- Python changes to `canup` subrepo
- Rework CMake to make shared files "opt-in" vs. "opt-out" across the
entire repo

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- Can update the binary on the H7 dev board on the bench
- CI

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).

---------

Co-authored-by: Gus Tahara-Edmonds <gtaharaedmonds@Guss-Laptop.local>
Co-authored-by: Gus Tahara-Edmonds <gtaharaedmonds@dhcp-206-12-41-148.ubcsecure.wireless.ubc.ca>