Refactoring fixes (#1139)
### Summary
<!-- Quick summary of changes, optional -->
A few minor changes to fix some things post-refactor.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Check for stack overflows at 1Hz instead of 1kHz on the DCM - 1kHz is
too fast and maxes out the CPU. The rest of the boards are 1Hz.
- Add `hw_assert.c` - includes our custom assert handler which breaks if
a debugger is connected.
- Fixed ordering of init for the DCM causes segfaults.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- Basic testing on every board on the car (not BMS).

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).