Fix ADC3 configuration on VC (#1176)
### Summary
<!-- Quick summary of changes, optional -->

We use 2 ADC3 channels on the VC. For multi-channel conversions on
STM32, you need to use DMA. However, the DMA controller used by ADC3
(BDMA), can only move to specific regions of memory (not normal RAM).

So, we need to find a workaround. The best solution I could find that
works reliably on the bench is to:
1. Configure ADC3 in single-channel interrupt mode.
2. After each conversions, re-configure the peripheral to read from the
other channel.

### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Changed ADC3 configuration as per above in Cube, and added support for
this in the VC's `hw_adc.c`.
- (see comment in `hw_adc.c` for a more in-depth explanation).

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

- Tested on bench, can read values from ADC3 IN0 and IN1 reliably.

### Resolved Issues
<!-- Link any issues that this PR resolved like so: `Resolves #1, #2,
and #5` (Note: Using this format, Github will automatically close the
issue(s) when this PR is merged in). -->

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).