Add load cells, and invoke io_wheels_inputCaptureCallback in Quadruna FSM (#1197)
### Summary
Add load cell app and io layers to the FSM, and invoke
`io_wheels_inputCaptureCallback` appropriately.

### Changelist 
- Added io and app layers for load cells to FSM.
- Added tests for load cell app layer.
- `HAL_TIM_IC_CaptureCallback` now defined in main.
- This invokes `tasks_timCaptureCallback`, which in turn invokes
`io_wheels_inputCaptureCallback`

### Note to Reviewer
On the Thruna FSM, `HAL_TIM_IC_CaptureCallback` was defined in an IO
layer to invoke the coolant and wheels input capture callbacks. It was
the only time we defined HAL layer functions anywhere outside of
`cubemx`. For consistency, we now define this function in `cubemx`, and
make it a wrapper for `tasks_timCaptureCallback`, which invokes our io
layer method, `io_wheels_inputCaptureCallback`.

### Testing Done
- Unit tests

### Resolved Issues

### Checklist
*Please change `[ ]` to `[x]` when you are ready.*
- [x] I have read and followed the code conventions detailed in
[README.md](../README.md) (*This will save time for both you and the
reviewer!*).
- [x] If this pull request is longer then **500** lines, I have provided
*explicit* justification in the summary above explaining why I *cannot*
break this up into multiple pull requests (*Small PR's are faster and
less painful for everyone involved!*).

---------

Co-authored-by: Edwin <20777515+Lucien950@users.noreply.github.com>