Unpacking CAN frames with JSONCAN in Python (#1271)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Add the ability to unpack raw CAN frames to signal values in JSONCAN
Python. This can be used with telemetry to decode received packets, or
used to pull logged frames from the SD card.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Was able to successfully decode CAN frames logged to the SD card, and
got expected values.