Specify data capture times for SD card and wireless telem in JSONCAN (#1273)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Add the ability to specify the data capture rate for both SD card
logging and wireless telemetry. The schema is as follows:
```json
"<msg name>": {
    <other signals omitted>
    "data_capture": {
        "log_cycle_time": 1000,
        "telem_cycle_time": 1000
     }
}
```

`log_cycle_time` is the cycle time at which the message will be logged
to the SD card. If `null`, this message will not be logged. This signal
can be omitted and the normal message cycle time will be used instead.

`telem_cycle_time` is the cycle time at which the message will be
transmitted via wireless telemetry. If `null`, this message will not be
sent. This signal can be omitted and the normal message cycle time will
be used instead.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

TODO