Quadruna CAN Charging (#1252)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

- Added dedicated BRUSA "board" for sending and receiving can msgs from
the charger.
- Added logic to have BMS control starting and stopping charging over
can.
- Added default values for charging currents, voltages and mains
currents
- Added translation functions from big to little endian (and vice versa)
and a function that uses it to convert floats for can msgs so that we
can set values directly in the code without having to do a manual
conversion to have the charger interpret the value we want.
- Added threshold of cycles we run after manual charging stop signal is
received and telling the charger to stop charging to stop the BMS from
immediately exiting charge state like we had seen testing with thruna
- Changed all (I think) io functions to do with charger to can
functions.
-Initialized all hbs from other boards to false on startup and all missing hb
faults to true - they will be cleared once other boards send a hb.


### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->
Tested charging with 120 and 240V at low and higher currents as well as tested CAN signals being sent and received from the BRUSA charger. Also added a unit test to ensure missing hb would not let the bms go into drive state when not connected to the other boards.

### Resolved Tickets
<!-- Link any tickets that this PR resolves. -->

---------

Co-authored-by: will-chaba <willchabaubc@gmail.com>