Simplify heartbeats + annotate alerts (#1282)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

A few changes that should make driving less fragile.

1. Changed the heartbeat dependency graph. Now, board A only checks the
heartbeat of board B if A uses a signal from board B. There currently
are a few instances of boards checking other boards' heartbeats, but
then not actually using signals from them. In my mind this is
unnecessary and could only lead to unnecessary faults down the line.
(check me on this!). This for example means the FSM and BMS checks no
other boards heartbeats.

2. Moved various faults on the VC related to the LV battery and currents
to warnings. If `AccumulatorRailOvercurrentFault` and
`BatteryRailOvercurrentFault` is set, we'd probably see total loss of
the LV bus, so no need to just exit drive state. If just one is set,
there is no immediate reason to stop driving. If there is a
`LvChargeFault` or `BoostControllerFault`, there is also no need to exit
drive state (we'd probably only notice the consequences of the failure
on the next power cycle). By making these faults instead of warnings we
reduce the chances of spurious shutdowns.

3. Added descriptions for all alerts.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Unit testing only.