Decode logfs CAN logs for Tracksight (#1295)
### Changelist 
<!-- Give a list of the changes covered in this PR. This will help both
you and the reviewer keep this PR within scope. -->

Added a script to decode raw log files written to the SD card. This uses
JSONCAN to unpack the signals and writes them to a CSV file. For how to
use, run:

```
python3 software/tracksight/backend/scripts/decode_raw_log.py --help
```

These CSV log files are (by default) written to
`software/tracksight/backend/data/`. Files in here are mounted to the
Docker backend when running in local mode, and so can be uploaded to the
local InfluxDB server for viewing on the Tracksight frontend. Pass the
paths to the unpacked CSV log files to `run_local.sh` and they will be
uploaded automatically.

Also, added a few CAN signals/warnings so we can check the state of
logging over CAN.

### Testing Done
<!-- Outline the testing that was done to demonstrate the changes are
solid. This could be unit tests, integration tests, testing on the car,
etc. Include relevant code snippets, screenshots, etc as needed. -->

Tested this whole pipeline on my laptop with VC #2.